feat: improve stdout process log output
